Reworking event delivery

+++ EventRouter/models/subscriptionContext.go
@@ -4,5 +4,5 @@ package models
 type SubscriptionContext struct {
 	Subscription Subscription `json:"subscription"`
 
-	Failures int `json:"failures"`
+	Failures uint16 `json:"failures"`
 }
